Added Normal Enemy Type, Player & Enemy Health UI, Improved Enemy collision.
In the first level obstacles were added which the enemy AI is able to navigate on. Another enemy type was imported in game with his own set of animations. The enemy now have a light to indicate where the player will take damage. The player and enemy now have a health bar to indicate their remaining health. Enemy health bar needs to be improved when there are multiple enemies in the same scene. The safe house is also improved with some minor designs in level 1.